Organized SearchHistory files
