+++ lib/screens/search_tab/components/search_history.dart
@@ -1,8 +1,8 @@
 import 'package:anonaddy/shared_components/list_tiles/alias_list_tile.dart';
 import 'package:anonaddy/shared_components/platform_aware_widgets/platform_scroll_bar.dart';
 import 'package:anonaddy/state_management/alias_state/fab_visibility_state.dart';
-import 'package:anonaddy/state_management/search/search_history_notifier.dart';
-import 'package:anonaddy/state_management/search/search_history_state.dart';
+import 'package:anonaddy/state_management/search/search_history/search_history_notifier.dart';
+import 'package:anonaddy/state_management/search/search_history/search_history_state.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
