+++ lib/state_management/search/search_history/search_history_notifier.dart
@@ -2,7 +2,7 @@ import 'dart:convert';
 
 import 'package:anonaddy/global_providers.dart';
 import 'package:anonaddy/models/alias/alias.dart';
-import 'package:anonaddy/state_management/search/search_history_state.dart';
+import 'package:anonaddy/state_management/search/search_history/search_history_state.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:hive/hive.dart';
